+++ src/schick/seg056.cpp
@@ -69,8 +69,8 @@ void buy_screen(void)
 	signed short l_di;
 	signed short j;
 	signed short items_x;
-	signed short fg_bak;
-	signed short bg_bak;
+	signed int fg_bak;
+	signed int bg_bak;
 	signed short l3;
 	signed short l4;
 	signed short percent;
